[refactor] : ConsoleView,Input 패키지 위치 수정 및, Main에 안쓰는 필드 삭제
